Modify Fragment to Activity
Add Activity in AndroidManifest
