--- app/src/main/java/com/example/once/Diary.kt
@@ -1,37 +1,0 @@
-package com.example.once
-
-import android.Manifest
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.net.Uri
-import android.os.Bundle
-import android.os.Environment
-import android.os.PersistableBundle
-import android.provider.MediaStore
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Button
-import android.widget.ImageView
-import androidx.appcompat.app.AlertDialog
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
-import java.io.File
-
-class Diary : Fragment() {
-
-    lateinit var uploadBtn: Button
-    lateinit var imgView: ImageView
-
-    private var REQUEST_READ_EXTERNAL_STORAGE = 1000
-
-    @Override
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-    }
-
-
-}
